+++ influxdb3/query_iterator.go
@@ -52,6 +52,8 @@ type QueryIterator struct {
 	reader *flight.Reader
 	// Current record
 	record arrow.Record
+	// The first err that might occur
+	err error
 	// Index of row of current object in current record
 	indexInRecord int
 	// Total index of current object
@@ -66,6 +68,7 @@ func newQueryIterator(reader *flight.Reader) *QueryIterator {
 	return &QueryIterator{
 		reader:        reader,
 		record:        nil,
+		err:           nil,
 		indexInRecord: -1,
 		i:             -1,
 		current:       nil,
@@ -84,6 +87,9 @@ func (i *QueryIterator) Next() bool {
 	i.i++
 	for i.record == nil || i.indexInRecord >= int(i.record.NumRows()) {
 		if !i.reader.Next() {
+			if readError := i.reader.Err(); readError != nil && i.err == nil {
+				i.err = i.reader.Err()
+			}
 			i.done = true
 			return false
 		}
@@ -187,6 +193,12 @@ func (i *QueryIterator) Index() interface{} {
 func (i *QueryIterator) Done() bool {
 	return i.done
 }
+
+// Err returns the first err that might have occurred during iteration
+//
+// Returns:
+//   - the err or nil if no err occurred
+func (i *QueryIterator) Err() error { return i.err }
 
 // Raw returns the underlying flight.Reader associated with the QueryIterator.
 // WARNING: It is imperative to use either the Raw method or the Value and Next functions, but not both at the same time,